+++ STACpopulator/populator_base.py
@@ -144,6 +144,14 @@ class STACpopulatorBase(ABC):
         # Add any assets if provided in the config
         self._collection_info["assets"] = self.__make_collection_assets()
 
+        # Cast providers to pystac objects
+        self._collection_info["providers"] = self.__make_collection_providers()
+
+        # Add contacts as extra_field
+        if "contacts" in self._collection_info:
+            self._collection_info["extra_fields"] = {}
+            self._collection_info["extra_fields"]["contacts"] = self.__make_collection_contacts()
+
         # Construct links if provided in the config. This needs to be done before constructing a collection object.
         collection_links = self.__make_collection_links()
 
@@ -179,6 +187,29 @@ class STACpopulatorBase(ABC):
             for asset_name, asset_info in self._collection_info["assets"].items():
                 pystac_assets[asset_name] = pystac.Asset(**asset_info)
         return pystac_assets
+
+    def __make_collection_providers(self) -> List[pystac.Provider]:
+        """Create collection level providers based on data read in from the configuration file.
+
+        :return: List of pystac Provider objects
+        :rtype: List[pystac.Provider]
+        """
+        pystac_providers = []
+        if "providers" in self._collection_info:
+            providers = self._collection_info.pop("providers")
+            pystac_providers = [pystac.Provider(**provider) for provider in providers]
+        return pystac_providers
+
+    def __make_collection_contacts(self) -> List[dict]:
+        """Create collection level contacts based on data read in from the configuration file.
+
+        :return: List of dictionnary contact objects
+        :rtype: List[dict]
+        """
+        contacts = []
+        if "contacts" in self._collection_info:
+            contacts = self._collection_info.pop("contacts")
+        return contacts
 
     def publish_stac_collection(self, collection_data: dict[str, Any]) -> None:
         """Publish this collection by uploading it to the STAC catalog at self.stac_host."""