+++ STACpopulator/extensions/base.py
@@ -340,7 +340,7 @@ class MetaItemExtension:
             key: asset
             for key, asset in self.item.get_assets().items()
             if (isinstance(service_type, ServiceType) and service_type.value in asset.extra_fields)
-            or any(ServiceType.from_value(field, default=None) is ServiceType for field in asset.extra_fields)
+            or any(ServiceType.from_value(field, default=False) for field in asset.extra_fields)
         }
 
     def __repr__(self) -> str: