+++ STACpopulator/extensions/file.py
@@ -75,10 +75,29 @@ class FileHelper(ExtensionHelper):
 
     @functools.cached_property
     def size(self) -> Optional[int]:
-        """Return file size in bytes, None if asset_key not in access URLs dictionnary."""
+        """Get file size in bytes.
+
+        Returns
+        -------
+            (int): File size in bytes or None if:
+                    - the `asset_key` is not in `access_urls` dict
+                    - the server does include a Content-Length header
+                    - the Content-Length value is non-numeric.
+        """
         if self.asset_key not in self.access_urls:
             logger.warning("Asset key %s is not present in access URLs.", self.asset_key)
             return None
         res = self.session.head(self.access_urls[self.asset_key])
         res.raise_for_status()
-        return int(res.headers.get("Content-Length", None))
+
+        content_length = res.headers.get("Content-Length")
+        try:
+            return int(content_length)
+        except (TypeError, ValueError) as e:
+            logger.exception(
+                "Failed to parse Content-Length header (%r) for asset key %s",
+                content_length,
+                self.asset_key,
+                exc_info=e,
+            )
+            return None