+++ CHANGES.md
@@ -4,8 +4,6 @@
 
 <!-- insert list items of new changes here -->
 
-## [1.12.0](https://github.com/crim-ca/stac-populator/tree/1.12.0) (2025-11-17)
-
 * Add `RDPS_CRIM` and `HRDPS_CRIM` implementations.
 * Add `cf` extension adding CF Parameter metadata to (H)RDPS stac asset and items.
 * Add `cf` and `file` helpers.
@@ -16,9 +14,16 @@
 * Update RDPS and HRDPS tests.
 * Add documentation for implementing new populators
 
-## [1.11.0](https://github.com/crim-ca/stac-populator/tree/1.11.0) (2025-11-12)
+## [0.11.0](https://github.com/crim-ca/stac-populator/tree/0.11.0) (2025-11-17)
 
 * Add option to automatically update collection extents and summaries based on ingested items.
+
+## [0.10.0](https://github.com/crim-ca/stac-populator/tree/0.10.0) (2025-11-11)
+
+* Add `pre-commit` linting rules (code format + STAC field sorting in JSON).
+* Add initial RDPS and HRDPS examples with minimal metadata from Ouranos THREDDS samples.
+* Fix `datacube` to report `[None, None]` instead of `["", ""]` for unbounded `cube:dimensions.extent`.
+* Fix unique ID creation from THREDDS access URLs.
 
 ## [0.9.0](https://github.com/crim-ca/stac-populator/tree/0.9.0) (2025-08-26)